@@ -1,12 +1,14 @@
 module github.com/FlowingSPDG/streamdeck-newtek
 
-go 1.17
-
-require github.com/FlowingSPDG/streamdeck v0.0.0-20221208184904-f88faf81846f
+go 1.19
 
 require (
-	github.com/FlowingSPDG/newtek-go v0.0.0-20221210114803-a47565f45bc9 // indirect
+	github.com/FlowingSPDG/newtek-go v0.0.0-20221210114803-a47565f45bc9
+	github.com/FlowingSPDG/streamdeck v0.0.0-20221211164639-19c5641697e8
+	github.com/oliamb/cutter v0.2.2
+)
+
+require (
 	github.com/gorilla/websocket v1.5.0 // indirect
 	github.com/icholy/digest v0.1.15 // indirect
-	github.com/oliamb/cutter v0.2.2 // indirect
 )
